refactor: separate I/O logic from main
